implement Perspective transformers
using Viewport3D and Viewport2DVisual3D. This renders the shapes on a (flat) 3D surface (quad), transforms it in 3D space and renders it with an OrthographicCamera. WPF doesn't support non-affine 2D transformations (2D transformation matrices only have 6 members).

This is completely untested and may crash!

It is loosely based on this article: http://www.charlespetzold.com/blog/2007/08/250638.html