set the correct size in ArrangeOverride in addition to the position
This is needed for the ViewPort3D shapes (i.e. the ones having perspective transformers) so they are (eventually) scaled correctly and independently of the overall control aspect ratio.

For Path based shapes this isn't needed, which is why it wasn't done before (lazyness). Paths are rendered at the top left of their viewport, so just setting the origin point was enough for the overall control to appear visually centered within its viewport. On the other hand, Viewport3D takes its available size into account to determine how big it should render itself.

The actual rendering of Viewport3D shapes is still not yet correct, so it doesn't work yet.